requirements.txt
################################################################################
# File: requirements.txt
# Project: Flask Bookstore API
# Description: List of dependencies required by the Flask application
# Author: [Satya Prakash Nigam <spnigam25@yahoo.com]
# Version: 1.0
# Date: [Nov 12, 2024]
################################################################################


# Flask Web Framework
Flask==3.0.3

# MySQL Database Connector
mysql-connector-python==9.1.0
mysqlclient==2.2.5

# Unit Testing Framework
pytest==8.3.3
pytest-cov==4.0.0
pytest-html==3.2.0

# Allure Framework for Test Reporting
allure-pytest==2.12.1

# Additional Dependencies
Werkzeug==3.1.3
atomicwrites==1.4.1
attrs==24.2.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
coverage==7.6.4
flasgger==0.9.7.1
Flask-MySQLdb==2.0.0
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
numpy==2.0.0  # Specific version included for pytest compatibility
packaging==24.2
pluggy==1.5.0
py==1.11.0
PyMySQL==1.1.1
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.16.0
SQLAlchemy==2.0.36
tomli==2.1.0
typing_extensions==4.12.2


################################################################################
# Installation Instructions:
#
# 1. Run the following command to install dependencies:
#    pip install -r requirements.txt
#
# 2. Verify installed packages:
#    pip freeze
################################################################################